Removed feature flag to enable pk sync Renamed and removed pk sync only jdbc tests bumped version in metadata and dockerfile update mysql change logs
