+++ airbyte-integrations/connectors/source-mysql-strict-encrypt/Dockerfile
@@ -24,6 +24,6 @@ ENV APPLICATION source-mysql-strict-encrypt
 
 COPY --from=build /airbyte /airbyte
 
-LABEL io.airbyte.version=3.0.9
+LABEL io.airbyte.version=3.1.0
 
 LABEL io.airbyte.name=airbyte/source-mysql-strict-encrypt
