+++ airbyte-integrations/connectors/source-mysql/Dockerfile
@@ -24,6 +24,6 @@ ENV APPLICATION source-mysql
 
 COPY --from=build /airbyte /airbyte
 
-LABEL io.airbyte.version=3.0.9
+LABEL io.airbyte.version=3.1.0
 
 LABEL io.airbyte.name=airbyte/source-mysql
